% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_aligned_transcript_annotation}
\alias{build_aligned_transcript_annotation}
\title{Build Aligned Transcript Column Annotation for ComplexHeatmap}
\usage{
build_aligned_transcript_annotation(anno_res, w_start, w_end)
}
\description{
Build Aligned Transcript Column Annotation for ComplexHeatmap
}
\keyword{internal}
