% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_coordinate_axis_annotation}
\alias{build_coordinate_axis_annotation}
\title{Build Genomic Coordinate Axis Bottom Annotation for ComplexHeatmap}
\usage{
build_coordinate_axis_annotation(w_start, w_end)
}
\description{
Build Genomic Coordinate Axis Bottom Annotation for ComplexHeatmap
}
\keyword{internal}
