% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_minimap}
\alias{build_minimap}
\title{Build the transcript/exon minimap}
\usage{
build_minimap(
  region_anno,
  tx_base,
  tx_exons,
  has_transcripts,
  chr,
  w_start,
  w_end
)
}
\description{
Build the transcript/exon minimap
}
\keyword{internal}
