% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{cat_palette}
\alias{cat_palette}
\title{Categorical palette for arbitrary numbers of levels}
\usage{
cat_palette(levels)
}
\description{
Categorical palette for arbitrary numbers of levels
}
\keyword{internal}
