% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{compute_heatmap_dimensions}
\alias{compute_heatmap_dimensions}
\title{Pixel height for the raster heatmap}
\usage{
compute_heatmap_dimensions(n_samples, n_facets)
}
\description{
Pixel height for the raster heatmap
}
\keyword{internal}
