% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{.bw_summary_one}
\alias{.bw_summary_one}
\title{Read one BigWig as a binned summary}
\usage{
.bw_summary_one(path, sname, gr, n_bins, type)
}
\description{
Read one BigWig as a binned summary
}
\keyword{internal}
