% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{attach_junction_positions}
\alias{attach_junction_positions}
\title{Attach y-positions and visual attrs to junctions with Interval Packing}
\usage{
attach_junction_positions(junctions, unique_samples, junc_band = 0.65)
}
\description{
Attach y-positions and visual attrs to junctions with Interval Packing
}
\keyword{internal}
