% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{build_context}
\alias{build_context}
\title{Assemble the app context (loaded data + worker pool)}
\usage{
build_context(
  se_dir,
  bigwig_base,
  anno_fst,
  bigwig_ext,
  n_workers,
  sj_se_dir = NULL
)
}
\description{
Assemble the app context (loaded data + worker pool)
}
\keyword{internal}
