% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{build_main_plot}
\alias{build_main_plot}
\title{Build the main wiggle ggplot}
\usage{
build_main_plot(
  plot_data,
  exon_highlights,
  junctions = NULL,
  bed_highlights = data.table::data.table(start = numeric(0), end = numeric(0)),
  bed_color = "#B22222",
  chr,
  w_start,
  w_end,
  overlap_factor,
  color_var,
  summary_type
)
}
\description{
Build the main wiggle ggplot
}
\keyword{internal}
