% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{build_plot_data}
\alias{build_plot_data}
\title{Join coverage with metadata, compute per-sample y-offsets and tooltips}
\usage{
build_plot_data(dt_full, meta_cur, facet_cols, overlap_factor, summary_type)
}
\description{
Join coverage with metadata, compute per-sample y-offsets and tooltips
}
\keyword{internal}
