% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_plot_reactive}
\alias{build_plot_reactive}
\title{Build the reactive graph that feeds the plot outputs}
\usage{
build_plot_reactive(input, ctx, rv, bed_data, timings_rv)
}
\description{
Build the reactive graph that feeds the plot outputs
}
\keyword{internal}
