% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{cat_palette}
\alias{cat_palette}
\title{Categorical palette for arbitrary numbers of levels}
\usage{
cat_palette(levels)
}
\description{
Concatenates several \code{pals} palettes to handle large categorical
variables (study_accession etc.) without recycling.
}
\keyword{internal}
