% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_minimap.R
\name{compute_plot_dimensions}
\alias{compute_plot_dimensions}
\title{Compute pixel heights for the main plot and minimap}
\usage{
compute_plot_dimensions(n_samples, n_facets, n_tx, has_color, minimap_override)
}
\description{
Compute pixel heights for the main plot and minimap
}
\keyword{internal}
