% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{downsample_by_study}
\alias{downsample_by_study}
\title{Cap samples per (study x facet-group) combination}
\usage{
downsample_by_study(meta, facet_cols, n)
}
\description{
Cap samples per (study x facet-group) combination
}
\keyword{internal}
