% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{format_ucsc_region}
\alias{format_ucsc_region}
\title{Format chr/start/end back into UCSC-style with thousands separators}
\usage{
format_ucsc_region(chr, start, end)
}
\description{
Format chr/start/end back into UCSC-style with thousands separators
}
\keyword{internal}
