% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{junction_palette}
\alias{junction_palette}
\title{Strand x annotation color palette for the junction layer}
\usage{
junction_palette()
}
\description{
Saturated = annotated, pale = novel. Blue = +, red = -, grey = \code{*}.
}
\keyword{internal}
