% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{parse_ucsc_region}
\alias{parse_ucsc_region}
\title{Parse a UCSC-style region string ("chr1:93,992,834-94,121,148")}
\usage{
parse_ucsc_region(s)
}
\description{
Parse a UCSC-style region string ("chr1:93,992,834-94,121,148")
}
\keyword{internal}
