% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{plot_region_heatmap}
\alias{plot_region_heatmap}
\title{Generate Heatmap Visualization of Read Coverage with Aligned Transcripts}
\usage{
plot_region_heatmap(
  ctx,
  chr,
  start,
  end,
  facet_cols = NULL,
  target_gene = NULL,
  max_samples = 3,
  min_expr = 5,
  bin_size = 0,
  scale_rows = TRUE,
  palette = "magma",
  show_transcripts = TRUE,
  show_row_names = FALSE,
  show_column_axis = TRUE,
  title = NULL,
  return_data = FALSE
)
}
\description{
Generate Heatmap Visualization of Read Coverage with Aligned Transcripts
}
