% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{read_region_bigwigs}
\alias{read_region_bigwigs}
\title{Read a genomic region across many BigWigs in parallel}
\usage{
read_region_bigwigs(
  samples,
  bw_file_map,
  chr,
  start,
  end,
  n_bins,
  summary_type,
  bp_backend
)
}
\description{
Read a genomic region across many BigWigs in parallel
}
\keyword{internal}
