% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{sanitize_metadata}
\alias{sanitize_metadata}
\title{Escape characters that would break ggiraph tooltips}
\usage{
sanitize_metadata(meta)
}
\description{
Escape characters that would break ggiraph tooltips
}
\keyword{internal}
