% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{subset_region_annotation}
\alias{subset_region_annotation}
\title{Subset annotation to features overlapping the window}
\usage{
subset_region_annotation(anno_dt, chr, w_start, w_end)
}
\description{
Subset annotation to features overlapping the window
}
\keyword{internal}
